+++ examples/hello/deps.nix
@@ -1,0 +1,10 @@
+let
+  nedryland = import ../../default.nix;
+  pkgs = import
+    (builtins.fetchTarball {
+      url = "https://github.com/NixOS/nixpkgs/archive/06278c77b5d162e62df170fec307e83f1812d94b.tar.gz";
+      sha256 = "sha256:11ri51840scvy9531rbz32241l7l81sa830s90wpzvv86v276aqs";
+    })
+    { };
+in
+(import ./project.nix { inherit nedryland pkgs; })