+++ overlays/backported-packages.nix
@@ -1,2 +1,0 @@
-# Remove packages from this file when we update the nixpkgs version to include them
-_self: _super: { }